requirements.txt
astroid==3.1.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
py4j==0.10.9.7
pylint==3.1.0
pyspark==3.5.1
pytest==8.1.1
python-dateutil==2.9.0.post0
ratelimit==2.2.1
requests==2.31.0
six==1.16.0
tomlkit==0.12.4
urllib3==2.2.1
